# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BusinessShield API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /websites/{id}/monitoring:
    get:
      operationId: getMonitoringConfig
      tags: [monitoring]
      summary: Get monitoring config for a website
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Monitoring config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoringConfig"
    put:
      operationId: upsertMonitoringConfig
      tags: [monitoring]
      summary: Create or update monitoring config
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitoringConfigInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoringConfig"

  # ─── Alerts ──────────────────────────────────────────────────────────────────
...
  /developer/keys:
    get:
      operationId: listApiKeys
      tags: [developer]
      summary: List API keys
      responses:
        "200":
          description: API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
    post:
      operationId: createApiKey
      tags: [developer]
      summary: Create a new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyBody"
      responses:
        "201":
          description: Created — key shown once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreated"
  /developer/keys/{id}:
    delete:
      operationId: revokeApiKey
      tags: [developer]
      summary: Revoke an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Revoked

  # ─── Organization & Onboarding ───────────────────────────────────────────────
...
  /onboarding/complete:
    post:
      operationId: completeOnboarding
      tags: [organizations]
      summary: Complete onboarding and save organization profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationProfile"
  /organizations/me:
    get:
      operationId: getMyOrganization
      tags: [organizations]
      summary: Get current user's organization profile
      responses:
        "200":
          description: Organization profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationProfile"
        "404":
          description: Not found
    patch:
      operationId: updateMyOrganization
      tags: [organizations]
      summary: Update organization profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationUpdate"
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationProfile"

  # ─── Audit Logs ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MonitoringConfig:
      type: object
      required: [id, websiteId, enabled, frequency, alertChannels, createdAt]
      properties:
        id:
          type: integer
        websiteId:
          type: integer
        enabled:
          type: boolean
        frequency:
          type: string
          enum: [hourly, daily, weekly]
        alertChannels:
          type: array
          items:
            type: string
        emailAlerts:
          type: boolean
        slackWebhookUrl:
          type: ["string", "null"]
        discordWebhookUrl:
          type: ["string", "null"]
        sslExpiryWarningDays:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    MonitoringConfigInput:
      type: object
      properties:
        enabled:
          type: boolean
        frequency:
          type: string
          enum: [hourly, daily, weekly]
        emailAlerts:
          type: boolean
        slackWebhookUrl:
          type: ["string", "null"]
        discordWebhookUrl:
          type: ["string", "null"]
        sslExpiryWarningDays:
          type: integer

    # ── Alerts ────────────────────────────────────────────────────────────────
...
    ApiKey:
      type: object
      required: [id, name, keyPrefix, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        keyPrefix:
          type: string
        isActive:
          type: boolean
        lastUsedAt:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ApiKeyCreated:
      type: object
      required: [id, name, keyPrefix, key, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        keyPrefix:
          type: string
        key:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    CreateApiKeyBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
        expiresAt:
          type: string

    # ── Organization & Onboarding ─────────────────────────────────────────────────
...
    OrganizationProfile:
      type: object
      required: [id, userId, name, industry, country, size, primaryConcerns, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        industry:
          type: string
          enum: [healthcare, fintech, ecommerce, saas, legal, media, education, nonprofit, other]
        country:
          type: string
        size:
          type: string
          enum: ["1-10", "11-50", "51-200", "200+", enterprise]
        primaryConcerns:
          type: array
          items:
            type: string
        websiteUrl:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        onboardingCompletedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    OnboardingInput:
      type: object
      required: [name, industry, country, size, primaryConcerns]
      properties:
        name:
          type: string
        industry:
          type: string
          enum: [healthcare, fintech, ecommerce, saas, legal, media, education, nonprofit, other]
        country:
          type: string
        size:
          type: string
          enum: ["1-10", "11-50", "51-200", "200+", enterprise]
        primaryConcerns:
          type: array
          items:
            type: string
        websiteUrl:
          type: string
        timezone:
          type: string
    OrganizationUpdate:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
          enum: [healthcare, fintech, ecommerce, saas, legal, media, education, nonprofit, other]
        country:
          type: string
        size:
          type: string
          enum: ["1-10", "11-50", "51-200", "200+", enterprise]
        primaryConcerns:
          type: array
          items:
            type: string
        websiteUrl:
          type: string
        timezone:
          type: string

    # ── Audit Logs ────────────────────────────────────────────────────────────────
...